async support doesn't require patching
