+++ tests/test_async.py
@@ -6,7 +6,7 @@ from jinja2 import ChainableUndefined
 from jinja2 import DictLoader
 from jinja2 import Environment
 from jinja2 import Template
-from jinja2.asyncsupport import auto_aiter
+from jinja2.async_utils import auto_aiter
 from jinja2.exceptions import TemplateNotFound
 from jinja2.exceptions import TemplatesNotFound
 from jinja2.exceptions import UndefinedError
